tools/requirements.txt
flask>=3.0.0
Pillow>=10.0.0
transformers>=4.40.0
torch>=2.2.0
